move param mapping stuff to the map class
